+++ src/main/java/roomescape/reservation/service/CreateReservationWithPaymentService.java
@@ -4,9 +4,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import roomescape.auth.service.dto.LoginMember;
 import roomescape.common.exception.InternalServerErrorException;
-import roomescape.infrastructure.TossPaymentClient;
-import roomescape.infrastructure.dto.request.ConfirmPaymentRequest;
-import roomescape.infrastructure.dto.response.ConfirmPaymentResponse;
+import roomescape.payment.infrastructure.TossPaymentClient;
+import roomescape.payment.infrastructure.dto.request.ConfirmPaymentRequest;
+import roomescape.payment.infrastructure.dto.response.ConfirmPaymentResponse;
 import roomescape.payment.service.PaymentService;
 import roomescape.reservation.service.dto.request.ReservationWithPaymentRequest;
 import roomescape.reservation.service.dto.response.ReservationWithPaymentResponse;
