+++ src/test/java/roomescape/reservation/service/CreateReservationWithPaymentServiceTest.java
@@ -14,9 +14,9 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import roomescape.auth.service.dto.LoginMember;
 import roomescape.common.exception.CustomException;
-import roomescape.infrastructure.TossPaymentClient;
-import roomescape.infrastructure.dto.request.ConfirmPaymentRequest;
-import roomescape.infrastructure.dto.response.ConfirmPaymentResponse;
+import roomescape.payment.infrastructure.TossPaymentClient;
+import roomescape.payment.infrastructure.dto.request.ConfirmPaymentRequest;
+import roomescape.payment.infrastructure.dto.response.ConfirmPaymentResponse;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.payment.domain.Payment;
