+++ src/test/java/roomescape/reservation/service/CreateReservationWithPaymentTest.java
@@ -13,11 +13,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import roomescape.auth.service.dto.LoginMember;
-import roomescape.infrastructure.TossPaymentClient;
+import roomescape.payment.infrastructure.TossPaymentClient;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
-import roomescape.infrastructure.dto.request.ConfirmPaymentRequest;
-import roomescape.infrastructure.dto.response.ConfirmPaymentResponse;
+import roomescape.payment.infrastructure.dto.request.ConfirmPaymentRequest;
+import roomescape.payment.infrastructure.dto.response.ConfirmPaymentResponse;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.service.dto.request.ReservationWithPaymentRequest;
